Access tokens now return old access tokens if they’re still current. Also added failsafe for race condition where multiple tokens quickly are generated by one user.

+++ fablersite/authentication/utils.py
@@ -2,9 +2,10 @@ from oauth2_provider.settings import oauth2_settings
 from oauthlib.common import generate_token
 from django.http import JsonResponse
 from oauth2_provider.models import AccessToken, Application, RefreshToken
-from django.utils.timezone import now, timedelta
- 
- 
+from django.utils.timezone import now, timedelta, datetime
+import pytz
+
+
 def get_token_json(access_token):
     """
     Takes an AccessToken instance as an argument
@@ -19,17 +20,17 @@ def get_token_json(access_token):
         'scope': access_token.scope
     }
     return JsonResponse(token)
- 
- 
+
+
 def get_access_token(user):
     """
     Takes a user instance and return an access_token as a JsonResponse
     instance.
     """
- 
+
     # our oauth2 app
     app = Application.objects.get(name="Fabler")
- 
+
     # We delete the old access_token and refresh_token
     try:
         old_access_token = AccessToken.objects.get(
@@ -38,34 +39,45 @@ def get_access_token(user):
             user=user, access_token=old_access_token
         )
     except:
-        pass
+        try:
+            old_access_tokens = AccessToken.objects.all().filter(user=user, application=app)
+            old_refresh_tokens = RefreshToken.objects.all().filter(user=user, access_token=old_access_token)
+        except:
+            # Something's wrong if we get here
+            pass
+        else:
+            old_access_tokens.delete()
+            old_refresh_tokens.delete()
     else:
+        timenow = datetime.utcnow().replace(tzinfo=pytz.utc)
+        if timenow < old_access_token.expires:
+            return get_token_json(old_access_token)
         old_access_token.delete()
         old_refresh_token.delete()
- 
+
     # we generate an access token
     token = generate_token()
     # we generate a refresh token
     refresh_token = generate_token()
- 
+
     expires = now() + timedelta(seconds=oauth2_settings.
                                 ACCESS_TOKEN_EXPIRE_SECONDS)
     scope = "read write"
- 
+
     # we create the access token
-    access_token = AccessToken.objects.\
+    access_token = AccessToken.objects. \
         create(user=user,
                application=app,
                expires=expires,
                token=token,
                scope=scope)
- 
+
     # we create the refresh token
-    RefreshToken.objects.\
+    RefreshToken.objects. \
         create(user=user,
                application=app,
                token=refresh_token,
                access_token=access_token)
- 
+
     # we call get_token_json and returns the access token as json
     return get_token_json(access_token)
